Rename Road Conditions generator/translator
