+++ Translators/RoadConditions/rc_tim_generator.py
